@@ -1,8 +1,10 @@
-const { SlashCommandBuilder } = require('discord.js')
+import { SlashCommandBuilder } from "discord.js"
 
-module.exports = {
-  data: new SlashCommandBuilder().setName('ping').setDescription('ping!!!!!!'),
+export const command = {
+  data: new SlashCommandBuilder()
+    .setName('ping')
+    .setDescription('ping!!!!'),
   async execute(interaction) {
-    await interaction.reply('pong!')
-  },
+    await interaction.reply('pong')
+  }
 }
